+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
-All notable changes to tool-NF-test.
+All notable changes to tool-NFTest.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).