+++ CHANGELOG.md
@@ -8,6 +8,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ---
 
 ## [Unreleased]
+### Changed
+- Wrap all NextFlow output with real-time line-by-line log statements
+- Add DEBUG lines about specific asserts
 
 ---