+++ CHANGELOG.md
@@ -8,6 +8,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ---
 
 ## [Unreleased]
+
+---
+
+## [1.0.0-rc.4] - 2023-10-04
 ### Changed
 - Wrap all NextFlow output with real-time line-by-line log statements
 - Wrap custom comparison to log through NFTest logger